Establish Reports namespace.  #3662.
This commit also moves CaseStatusReport into said namespace.
