--- lib/reporting.rb
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-module Reporting
-  autoload :CaseStatusReport, 'reporting/case_status_report'
-end
